System Templates/Remodel-Template/requirements.txt
# Minimal isolated env for template scaffold
python-dotenv>=1.0.1
pydantic>=2.6
pytest>=8.2
psycopg2-binary>=2.9.9
fastapi>=0.112.0
uvicorn>=0.30.0
